pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "whisperkey-mac"
version = "0.2.0"
description = "Voice input for macOS — hold a key to speak, release to transcribe"
requires-python = ">=3.10"
dependencies = [
    "faster-whisper>=1.0",
    "pynput>=1.7",
    "sounddevice>=0.4",
    "soundfile>=0.12",
    "numpy>=1.24",
    "pyperclip>=1.8",
    "opencc-python-reimplemented>=0.1",
    "rich>=13.0",
    "pyobjc-framework-Cocoa>=10.0",
    "pyobjc-framework-ApplicationServices>=10.0",  # AX API for DET-01/DET-02
]

[project.scripts]
whisperkey = "whisperkey_mac.main:main"

[tool.setuptools.packages.find]
where = ["."]
